[fix]: DDP evaluation hanging
Distributed validation on DDP must be done with local model
replicas, using `.module()`, instead of the DDP wrapper, to prevent
models from attempting to create buckets and synchronize, hanging
as the result.

Added a ternary operator to the `_forward()` to call it during
validation, and to call the DDP wrapper during training.